+++ tests/030_manifest.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
+import io
 import json
+import os
 import pytest
 
 from py_css_styleguide.manifest import Manifest
 
 
-def test_manifest_load():
+def test_manifest_load_string():
     source = (
         '.styleguide-metas-references{\n'
         '    names: "palette schemes";\n'
@@ -26,6 +28,43 @@ def test_manifest_load():
 
     manifest = Manifest()
     manifest.load(source)
+
+    assert manifest._path == None
+
+    assert sorted(manifest.metas.get('references')) == sorted([
+        'palette',
+        'schemes'
+    ])
+
+    assert manifest.palette == {
+        'white': '#ffffff',
+        'black': '#000000',
+    }
+
+    assert manifest.schemes == {
+        'black': {
+            'selectors': ".bg-black",
+            'values': "#000000",
+        },
+        'white': {
+            'values': "#ffffff",
+            'selectors': ".bg-white",
+        },
+    }
+
+
+def test_manifest_load_fileobject(fixtures_settings):
+    source_filepath = os.path.join(
+        fixtures_settings.fixtures_path,
+        "manifest_sample.css"
+    )
+
+    manifest = Manifest()
+
+    with io.open(source_filepath, 'r') as fp:
+        manifest.load(fp)
+
+    assert manifest._path == source_filepath
 
     assert sorted(manifest.metas.get('references')) == sorted([
         'palette',